chromosomePlots {Clonality}	R Documentation

Per-chromosome plots of the copy number arrays from a particular patient

Description

The function produces a sequence of plots for each chromosome with one-step segmented data of all samples of a particular patient.

Usage

chromosomePlots(data.seg1, ptlist, ptname,nmad)

Arguments

data.seg1
Output of one-step segmentation - output OneStepSeg of clonality.analysis().

ptlist
Vector of the patient IDs in the order of the samples appearing in the data. For example, if the first three tumors belong to patient A, and the following two belong to patient B, then ptlist=c('ptA', 'ptA', 'ptA', 'ptB', 'ptB').

ptname
Name of the patient from ptlist for which the data should be plotted

nmad
Number of MADs (median absolute deviations) that is used for Gain/Loss calls. Used to mark the Gain/Loss threshold on the plots.

Details

The function produces a sequence of plots for each chromosome with one-step segmented data of all samples of a particular patient. The dotted horizontal lines denote the gain and loss thresholds.

Examples

 # See example as in clonality.analysis()
   

[Package Clonality version 1.16.0 Index]